drop user, products -> product, stock -> available + img_data

+++ src/main/java/com/ecommerce/backend/repository/ProductRepository.java
@@ -4,10 +4,11 @@ import com.ecommerce.backend.models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
     boolean existsProductByName(String name);
 
-    boolean existsProductById(Integer id);
+    boolean existsProductById(Long id);
 
 }
